chore: use tempfile as context manager even when not deleting it
Use context manager to write out content, but keep it around for the
filename and delete that later.

Discovered by ruff: https://docs.astral.sh/ruff/rules/open-file-with-context-handler/

+++ weblate/formats/tests/test_convert.py
@@ -43,14 +43,13 @@ class ConvertFormatTest(BaseFormatTest):
     def test_convert(self) -> None:
         if not self.CONVERT_TEMPLATE:
             raise SkipTest(f"Test template not provided for {self.FORMAT.format_id}")
-        template = NamedTemporaryFile(delete=False, mode="w+")
-        translation = NamedTemporaryFile(delete=False, mode="w+")
+        translation = template = None
         try:
             # Generate test files
-            template.write(self.CONVERT_TEMPLATE)
-            template.close()
-            translation.write(self.CONVERT_TRANSLATION)
-            translation.close()
+            with NamedTemporaryFile(delete=False, mode="w+") as template:
+                template.write(self.CONVERT_TEMPLATE)
+            with NamedTemporaryFile(delete=False, mode="w+") as translation:
+                translation.write(self.CONVERT_TRANSLATION)
 
             # Parse
             storage = self.FORMAT(
@@ -78,8 +77,10 @@ class ConvertFormatTest(BaseFormatTest):
             with open(translation.name) as handle:
                 self.assertEqual(handle.read(), self.CONVERT_EXPECTED)
         finally:
-            os.unlink(template.name)
-            os.unlink(translation.name)
+            if template:
+                os.unlink(template.name)
+            if translation:
+                os.unlink(translation.name)
 
 
 class HTMLFormatTest(ConvertFormatTest):
